refactor: 컨벤션 등 수정
