+++ server/src/main/java/com/ahmadda/domain/EventPokeRepository.java
@@ -6,8 +6,8 @@ import java.time.LocalDateTime;
 
 public interface EventPokeRepository extends JpaRepository<EventPoke, Long> {
 
-    //TODO 성능 매우 나쁨!! 반드시 개선 필요
-    //TODO. 우선 빠르게 구현후 인덱싱 혹은 메모리 방식으로 전환 논의 필요
+    //TODO 성능 매우 나쁨!! 개선 필요
+    //TODO. 우선 빠르게 구현후 인덱싱 혹은 메모리 방식으로 전환 논의 필요후
     int countEventPokeHistoryByEventAndSenderAndRecipientAndSentAtAfter(
             final Event event,
             final OrganizationMember sender,
